+++ packages/veui/demo/cases/Checkbox.vue
@@ -14,26 +14,26 @@
     </veui-checkbox> {{ current }}
   </section>
   <section>
-    <veui-checkbox v-model="picked1">
+    <veui-checkbox v-model="checked1">
       正常状态
     </veui-checkbox>
   </section>
   <section>
     <veui-checkbox
-      v-model="picked2"
+      v-model="checked2"
       ui="small"
     >
       正常状态
     </veui-checkbox>
   </section>
   <section>
-    <veui-checkbox v-model="picked3">
+    <veui-checkbox v-model="checked3">
       选中状态
     </veui-checkbox>
   </section>
   <section>
     <veui-checkbox
-      v-model="picked4"
+      v-model="checked4"
       ui="small"
     >
       选中状态
@@ -41,7 +41,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked5"
+      v-model="checked5"
       :indeterminate="indeterminate5"
       @change="indeterminate5 = false"
     >
@@ -50,7 +50,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked6"
+      v-model="checked6"
       ui="small"
       :indeterminate="indeterminate6"
       @change="indeterminate6 = false"
@@ -60,7 +60,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked7"
+      v-model="checked7"
       disabled
     >
       选中无效状态
@@ -68,7 +68,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked8"
+      v-model="checked8"
       ui="small"
       disabled
     >
@@ -77,7 +77,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked9"
+      v-model="checked9"
       disabled
     >
       未选无效状态
@@ -85,7 +85,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked10"
+      v-model="checked10"
       ui="small"
       disabled
     >
@@ -94,7 +94,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked11"
+      v-model="checked11"
       disabled
       indeterminate
     >
@@ -103,7 +103,7 @@
   </section>
   <section>
     <veui-checkbox
-      v-model="picked12"
+      v-model="checked12"
       ui="small"
       disabled
       indeterminate
@@ -118,36 +118,62 @@
     >
       外部控制 <code>checked</code>
     </veui-checkbox>
+    <veui-button ui="tiny" @click="checked13 = !checked13">Toggle</veui-button>
+  </section>
+  <section>
+    <veui-checkbox :checked.sync="checked14">
+      <code>:checked.sync</code> 绑定状态：{{ checked14 }}
+    </veui-checkbox>
+    <veui-button ui="tiny" @click="checked14 = !checked14">Toggle</veui-button>
+  </section>
+  <section>
+    <h4>复选框组</h4>
+    <p>
+      <veui-checkbox
+        :checked="group.length > 0"
+        :indeterminate="group.length > 0 && group.length < 3"
+        @change="group.length ? group = [] : group = ['A', 'B', 'C']"
+      >全部</veui-checkbox>
+    </p>
+    <p>
+      <veui-checkbox v-model="group" value="A">A</veui-checkbox>
+      <veui-checkbox v-model="group" value="B">B</veui-checkbox>
+      <veui-checkbox v-model="group" value="C">C</veui-checkbox>
+    </p>
+    <p>{{ group }}</p>
   </section>
 </article>
 </template>
 
 <script>
-import { Checkbox } from 'veui'
+import { Checkbox, Button } from 'veui'
 
 export default {
   name: 'checkbox-demo',
   components: {
-    'veui-checkbox': Checkbox
+    'veui-checkbox': Checkbox,
+    'veui-button': Button
   },
   data () {
     return {
       current: '已选',
-      picked1: false,
-      picked2: false,
-      picked3: true,
-      picked4: true,
-      picked5: true,
+      checked1: false,
+      checked2: false,
+      checked3: true,
+      checked4: true,
+      checked5: true,
       indeterminate5: true,
-      picked6: false,
+      checked6: false,
       indeterminate6: true,
-      picked7: true,
-      picked8: true,
-      picked9: false,
-      picked10: false,
-      picked11: false,
-      picked12: false,
-      checked13: true
+      checked7: true,
+      checked8: true,
+      checked9: false,
+      checked10: false,
+      checked11: false,
+      checked12: false,
+      checked13: true,
+      checked14: false,
+      group: []
     }
   }
 }
@@ -156,5 +182,10 @@ export default {
 <style lang="less" scoped>
 section {
   margin-bottom: 1em;
+}
+
+.veui-checkbox + .veui-button,
+.veui-checkbox + .veui-checkbox {
+  margin-left: 20px;
 }
 </style>