+++ packages/veui/demo/cases/Radio.vue
@@ -85,6 +85,15 @@
       未选禁用状态
     </veui-radio>
   </section>
+  <section>
+    <h3>单选组</h3>
+    <p>
+      <veui-radio v-model="group" value="A">A</veui-radio>
+      <veui-radio v-model="group" value="B">B</veui-radio>
+      <veui-radio v-model="group" value="C">C</veui-radio>
+    </p>
+    <p>{{ group }}</p>
+  </section>
 </article>
 </template>
 
@@ -100,7 +109,8 @@ export default {
   data () {
     return {
       size: null,
-      checked: false
+      checked: false,
+      group: null
     }
   }
 }